sanitise sign comparisons
This is a mechanical change addressing the various sign comparison warnings that
are identified by both clang and gcc.  This helps cleanup some of the warning
spew that occurs during builds.

llvm-svn: 205390

+++ lldb/source/Commands/CommandObjectBreakpoint.cpp
@@ -1843,7 +1843,7 @@ CommandObjectMultiwordBreakpoint::VerifyBreakpointIDs (Args &args, Target *targe
             if (breakpoint != NULL)
             {
                 const size_t num_locations = breakpoint->GetNumLocations();
-                if (cur_bp_id.GetLocationID() > num_locations)
+                if (static_cast<size_t>(cur_bp_id.GetLocationID()) > num_locations)
                 {
                     StreamString id_str;
                     BreakpointID::GetCanonicalReference (&id_str,
