+++ lldb/source/Core/Opcode.cpp
@@ -63,7 +63,7 @@ Opcode::Dump (Stream *s, uint32_t min_byte_width)
 
     // Add spaces to make sure bytes dispay comes out even in case opcodes
     // aren't all the same size
-    if (bytes_written < min_byte_width)
+    if (static_cast<uint32_t>(bytes_written) < min_byte_width)
         bytes_written = s->Printf ("%*s", min_byte_width - bytes_written, "");
     return bytes_written;
 }
