+++ lldb/source/DataFormatters/CF.cpp
@@ -160,7 +160,7 @@ lldb_private::formatters::CFBitVectorSummaryProvider (ValueObject& valobj, Strea
     if (error.Fail() || num_bytes == 0)
         return false;
     uint8_t *bytes = buffer_sp->GetBytes();
-    for (int byte_idx = 0; byte_idx < num_bytes-1; byte_idx++)
+    for (uint64_t byte_idx = 0; byte_idx < num_bytes-1; byte_idx++)
     {
         uint8_t byte = bytes[byte_idx];
         bool bit0 = (byte & 1) == 1;
